+++ akka-http-tests/src/test/java/akka/http/javadsl/server/directives/PathDirectivesTest.java
@@ -375,4 +375,44 @@ public class PathDirectivesTest extends JUnitRouteTest {
          .assertStatusCode(200)
          .assertEntity("Ok");
   }
+
+  @Test
+  public void testIgnoreTrailingSlash() {
+    TestRoute route = testRoute(
+      ignoreTrailingSlash(() ->
+        route(
+          path("foo", () -> complete("Ok")),
+          path("bar", () -> pathEndOrSingleSlash(() -> complete("Ok"))),
+          path(PathMatchers.segment("baz").slash(), () -> complete("Ok"))
+        )
+      )
+    );
+
+    route.run(HttpRequest.GET("/foo"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/foo/"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/bar/"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/baz"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/baz/"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/foo/?query#frag"))
+      .assertStatusCode(StatusCodes.OK)
+      .assertEntity("Ok");
+
+    route.run(HttpRequest.GET("/foz/"))
+      .assertStatusCode(StatusCodes.NOT_FOUND);
+  }
 }