+++ akka-http-tests/src/test/scala/akka/http/scaladsl/server/directives/PathDirectivesSpec.scala
@@ -4,6 +4,8 @@
 
 package akka.http.scaladsl.server.directives
 
+import java.util.concurrent.atomic.AtomicInteger
+
 import scala.collection.immutable.ListMap
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server._
@@ -473,6 +475,63 @@ class PathDirectivesSpec extends RoutingSpec with Inside {
       Get("/foo/bar/") ~>
         redirectToNoTrailingSlashIfPresent(MovedPermanently) { completeOk } ~>
         check { status shouldEqual MovedPermanently }
+    }
+  }
+
+  "the `ignoreTrailingSlash` directive" should {
+    def route(counter: AtomicInteger = new AtomicInteger(0)) = ignoreTrailingSlash {
+      counter.incrementAndGet()
+      path("foo") {
+        complete(s"${counter.get()}")
+      } ~
+        (pathPrefix("bar") & pathEndOrSingleSlash) {
+          complete(s"${counter.get()}")
+        } ~
+        path("baz" /) {
+          complete(s"${counter.get()}")
+        }
+    }
+
+    "pass if the request path doesn't have a trailing slash" in {
+      Get("/foo") ~> route() ~> check {
+        responseAs[String] shouldEqual "1"
+      }
+    }
+
+    "pass if the request path has a trailing slash by retrying the inner route" in {
+      Get("/foo/") ~> route() ~> check {
+        responseAs[String] shouldEqual "2"
+      }
+    }
+
+    "pass when the request contains parameters and fragments" in {
+      Get("/foo/?query#frag") ~> route() ~> check {
+        responseAs[String] shouldEqual "2"
+      }
+    }
+
+    "retry the inner route only once if path already checks for an optional trailing slash" in {
+      Get("/bar/") ~> route() ~> check {
+        responseAs[String] shouldEqual "1"
+      }
+    }
+
+    "retry accordingly if the path expects a trailing slash" in {
+      Get("/baz") ~> route() ~> check {
+        responseAs[String] shouldEqual "2"
+      }
+
+      Get("/baz/") ~> route() ~> check {
+        responseAs[String] shouldEqual "1"
+      }
+    }
+
+    "reject if request can't be matched with nor without a trailing slash" in {
+      val counter = new AtomicInteger(0)
+      Get("/foz") ~> route(counter) ~> check {
+        counter.get() shouldEqual 2
+        handled shouldEqual false
+      }
     }
   }